cmd/sniproxy: add APPC_SSH_DEST environment variable.
sniproxy is packaged in a container using
cgr.dev/chainguard/static:latest as the base image, which
has no shell. Not having a shell is nice from a security
perspective, but it makes it hard to turn environment
variables passed to the container into arguments to sniproxy.

Add support for APPC_SSH_DEST, a FQDN environment variable
to which the SSH port should be forwarded.

Updates https://github.com/tailscale/tailscale/issues/1748

Signed-off-by: Denton Gentry <dgentry@tailscale.com>